Harmonic Balance Works!

+++ ROUTINES/HARMONIC/TIMESERIES_DERIV.m
@@ -116,7 +116,7 @@ function [x_t, h] = TIMESERIES_DERIV(Nt,h,X0,ord)
     if size(Xf,1)~=Nt
         disp('error');
     end
-    Xf = Xf*Nt/2;
+    Xf = Xf*single(Nt/2);
     x_t = real(ifft(Xf));    
 end
